// src/enums/PersonalData.ts
export enum PersonalData {
    ORDER_SURNAME = 'ORDER_SURNAME',
    ORDER_NAME = 'ORDER_NAME',
    ORDER_OWNERSHIP = 'ORDER_OWNERSHIP',
    ORDER_PATRONYMIC = 'ORDER_PATRONYMIC',
    ORDER_PHONE = 'ORDER_PHONE',
    ORDER_EMAIL = 'ORDER_EMAIL',
    ORDER_CITY = 'ORDER_CITY',
    ORDER_ADDRESS = 'ORDER_ADDRESS',
}